# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: User & Task API
  description: API для работы с пользователями и задачами
  version: "1.0"
paths:
...
  /users:
    get:
      summary: Get all users
      operationId: GetUsers
      tags:
        - users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      summary: Create new user
      operationId: PostUser
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User successfully created
  /users/{id}:
    get:
      summary: Get user by ID
      operationId: GetUserByID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    patch:
      summary: Update user by ID
      operationId: PatchUserByID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated
        '404':
          description: User not found

    delete:
      summary: Delete user by ID
      operationId: DeleteUserByID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found

  # Endpoints for Tasks
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: "password123"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
          description: List of tasks assigned to the user
    Task:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the task
        task:
          type: string
          description: Description of the task
        is_done:
          type: boolean
          description: Indicates whether the task is completed
        user_id:
          type: integer
          description: ID of the user associated with the task
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the task was soft-deleted
...